requirements.txt
antlr4-python3-runtime==4.13.2
typing-extensions==4.12.2
treelib==1.7.0
numpy==2.1.3
networkx==3.4.2